Merge in SEQ/pa-mongo from bugfix/PTSD-764-traceviewing-kestrel-data-off-by-16-frames to develop

* commit 'c4ad039f02da3c92f50360cbc89ec3f8455469ec':
  PTSD-764. Properly fix TestPulseAccumulator unit test by starting test at frames [512,1024).
  PTSD-764. Some merge fixes.
  PTSD-764. Fix unit test.
  PTSD-764. Allow for frame indices to be signed.
  PTSD-764. Address reviewer comments.
  PTSD-764. Account for viterbi latency by adjusting label metadata frame boundaries.

+++ src/basecaller/basecaller/test/TestBaseliner.cpp
@@ -288,7 +288,7 @@ TEST_P(MultiScaleBaseliner, StatsAndSubtraction)
          auto& baseliner = *baseliners[meta.PoolId()];
          auto cameraBatch = baseliner(batch);
 
-         if (batch.Metadata().FirstFrame() < burnInFrames) continue;
+         if (batch.Metadata().FirstFrame() < static_cast<int32_t>(burnInFrames)) continue;
 
          auto traces = std::move(cameraBatch.first);
          auto stats = std::move(cameraBatch.second);
